AI-Devs-3-Solutions-App/requirements.txt
streamlit
openai
python-dotenv
beautifulsoup4
requests